refactor: split main.ts into rankings-view and city-detail-view
Closes #158
