+++ src/frontend/main.ts
@@ -1,199 +1,37 @@
+/**
+ * Main orchestrator for ETS2 Trucker Advisor
+ *
+ * Handles initialization, routing between rankings and city detail views,
+ * DLC banner, onboarding section, and How It Works toggle.
+ * All rendering is delegated to rankings-view.ts and city-detail-view.ts.
+ */
+
 import { initPageData } from './page-init.js';
-import { computeFleetAsync, computeRankingsAsync } from './optimizer-client.js';
-import type { CityRanking, FleetEntry, OptimalFleetEntry } from './optimizer.js';
 import {
-  getOwnedGarages, toggleOwnedGarage, isOwnedGarage,
-  getFilterMode, setFilterMode,
-  getSelectedCountries, setSelectedCountries,
   isFirstVisit, isBannerDismissed, dismissBanner,
   isOnboardingCollapsed, setOnboardingCollapsed,
 } from './storage.js';
-import { copyToClipboard } from './clipboard.js';
-import { normalize } from './data.js';
 import type { AllData, Lookups } from './data.js';
+import type { CityRanking } from './optimizer.js';
+import {
+  renderRankings, initRankingsView,
+  showLoading, showError,
+  type RankingsState,
+} from './rankings-view.js';
+import { renderCity } from './city-detail-view.js';
 
-let data: AllData | null = null;
-let lookups: Lookups | null = null;
+// ============================================
+// Shared state
+// ============================================
+
+const state: RankingsState = {
+  data: null as unknown as AllData,
+  lookups: null as unknown as Lookups,
+  cachedRankings: null as CityRanking[] | null,
+  displayedRankings: null as CityRanking[] | null,
+};
+
 let currentCityId: string | null = null;
-let cachedRankings: CityRanking[] | null = null;
-let displayedRankings: CityRanking[] | null = null;
-
-function debounce(fn: () => void, ms: number): () => void {
-  let timer: ReturnType<typeof setTimeout>;
-  return () => {
-    clearTimeout(timer);
-    timer = setTimeout(fn, ms);
-  };
-}
-
-function formatNumber(n: number): string {
-  return Math.round(n).toLocaleString();
-}
-
-function getUniqueCountries(): string[] {
-  if (!data || !data.cities) return [];
-  const countries = Array.from(new Set(data.cities.map((c) => c.country)));
-  return countries.sort();
-}
-
-// ============================================
-// Country filter dropdown
-// ============================================
-
-function toggleDropdown() {
-  const dropdown = document.getElementById('country-dropdown')!;
-  const btn = document.getElementById('country-filter-btn')!;
-  const isVisible = dropdown.style.display !== 'none';
-  if (isVisible) {
-    dropdown.style.display = 'none';
-    btn.setAttribute('aria-expanded', 'false');
-  } else {
-    dropdown.style.display = 'block';
-    btn.setAttribute('aria-expanded', 'true');
-    const firstCheckbox = dropdown.querySelector('input[type="checkbox"]');
-    if (firstCheckbox) (firstCheckbox as HTMLElement).focus();
-  }
-}
-
-function closeDropdown() {
-  const dropdown = document.getElementById('country-dropdown')!;
-  const btn = document.getElementById('country-filter-btn')!;
-  dropdown.style.display = 'none';
-  btn.setAttribute('aria-expanded', 'false');
-}
-
-function updateCountryButtonText() {
-  const selected = getSelectedCountries();
-  const btn = document.getElementById('country-filter-btn')!;
-  if (selected.length === 0) {
-    btn.textContent = 'All Countries';
-    btn.setAttribute('aria-label', 'Filter by country');
-  } else if (selected.length === 1) {
-    btn.textContent = '1 Country';
-    btn.setAttribute('aria-label', 'Filter by country, 1 selected');
-  } else {
-    btn.textContent = `${selected.length} Countries`;
-    btn.setAttribute('aria-label', `Filter by country, ${selected.length} selected`);
-  }
-}
-
-function renderCountryCheckboxes() {
-  const countries = getUniqueCountries();
-  const countryOptions = document.getElementById('country-options')!;
-  const selected = getSelectedCountries();
-
-  countryOptions.innerHTML = `
-    <label class="country-option all-countries" role="option">
-      <input type="checkbox" id="all-countries-checkbox"
-        aria-checked="${selected.length === 0 ? 'true' : 'false'}"
-        ${selected.length === 0 ? 'checked' : ''}>
-      <span>All Countries</span>
-    </label>
-    ${countries.map((country) => `
-      <label class="country-option" role="option">
-        <input type="checkbox" value="${country}"
-          aria-checked="${selected.includes(country) ? 'true' : 'false'}"
-          aria-label="${country}"
-          ${selected.includes(country) ? 'checked' : ''}>
-        <span>${country}</span>
-      </label>
-    `).join('')}
-  `;
-
-  document.getElementById('all-countries-checkbox')!.addEventListener('change', (e) => {
-    if ((e.target as HTMLInputElement).checked) {
-      setSelectedCountries([]);
-      renderCountryCheckboxes();
-      updateCountryButtonText();
-      renderRankings();
-    }
-  });
-
-  countryOptions.querySelectorAll('input[type="checkbox"]:not(#all-countries-checkbox)').forEach((checkbox) => {
-    checkbox.addEventListener('change', (e) => {
-      const country = (e.target as HTMLInputElement).value;
-      const sel = getSelectedCountries();
-      if ((e.target as HTMLInputElement).checked) {
-        if (!sel.includes(country)) setSelectedCountries([...sel, country]);
-      } else {
-        setSelectedCountries(sel.filter((c) => c !== country));
-      }
-      renderCountryCheckboxes();
-      updateCountryButtonText();
-      renderRankings();
-    });
-  });
-}
-
-// ============================================
-// Garage count badge
-// ============================================
-
-function updateGarageCount() {
-  const ownedGarages = getOwnedGarages();
-  if (!data || !lookups) {
-    document.getElementById('garage-count')!.textContent = ownedGarages.length.toString();
-    return;
-  }
-  const searchTerm = normalize(citySearch.value);
-  const selectedCountries = getSelectedCountries();
-  let count = 0;
-  for (const cityIdStr of ownedGarages) {
-    const city = lookups.citiesById.get(cityIdStr);
-    if (!city) continue;
-    if (searchTerm && !normalize(city.name).includes(searchTerm) && !normalize(city.country).includes(searchTerm)) continue;
-    if (selectedCountries.length > 0 && !selectedCountries.includes(city.country)) continue;
-    count++;
-  }
-  document.getElementById('garage-count')!.textContent = count.toString();
-}
-
-// ============================================
-// Results count feedback
-// ============================================
-
-function updateResultsCount(shown: number, total: number) {
-  if (shown === total || total === 0) {
-    resultsCount.textContent = '';
-  } else {
-    resultsCount.textContent = `Showing ${shown} of ${total} cities`;
-  }
-}
-
-// ============================================
-// Score tier helpers
-// ============================================
-
-interface ScoreTier {
-  className: string;
-  label: string;
-}
-
-function getScoreTier(index: number, total: number): ScoreTier {
-  if (total === 0) return { className: '', label: '' };
-  const percentile = (index / total) * 100;
-  if (percentile < 10) return { className: 'score-tier-excellent', label: 'Excellent — top 10%' };
-  if (percentile < 25) return { className: 'score-tier-good', label: 'Good — top 25%' };
-  if (percentile < 50) return { className: 'score-tier-average', label: 'Average — top 50%' };
-  return { className: 'score-tier-below', label: 'Below average — bottom 50%' };
-}
-
-// ============================================
-// Rank helpers
-// ============================================
-
-function getCityRank(cityId: string) {
-  if (!displayedRankings) return null;
-  const index = displayedRankings.findIndex((r) => r.id === cityId);
-  if (index === -1) return null;
-  return { rank: index + 1, total: displayedRankings.length };
-}
-
-function formatRank(rank: number, total: number): string {
-  const isTopTier = rank <= Math.ceil(total * 0.1);
-  const className = isTopTier ? 'rank-display top-tier' : 'rank-display';
-  return `<span class="${className}"><span class="rank">#${rank}</span> of ${total}</span>`;
-}
 
 // ============================================
 // DOM elements
@@ -211,450 +49,17 @@ const howItWorksToggle = document.getElementById('how-it-works-toggle');
 const onboardingToggle = document.getElementById('onboarding-toggle');
 const onboardingSection = document.getElementById('onboarding');
 
-// Filter toggle
-filterToggle.addEventListener('click', (e) => {
-  const btn = (e.target as HTMLElement).closest('.filter-btn');
-  if (!btn) return;
-  const mode = btn.getAttribute('data-filter')!;
-  filterToggle.querySelectorAll('.filter-btn').forEach((b) => b.classList.remove('active'));
-  btn.classList.add('active');
-  setFilterMode(mode);
-  renderRankings();
-});
-
 // ============================================
-// Rankings rendering
+// Utilities
 // ============================================
 
-function summarizeTrailers(fleet: FleetEntry[]): string {
-  return fleet.map(e => e.displayName).join(', ');
-}
-
-async function renderRankings() {
-  const rankings = await computeRankingsAsync(data!, lookups!);
-  cachedRankings = rankings;
-
-  if (rankings.length === 0) {
-    cachedRankings = null;
-    rankingsContent.innerHTML = '<div class="empty-state">No cities with data yet.</div>';
-    updateResultsCount(0, 0);
-    return;
-  }
-
-  const searchTerm = normalize(citySearch.value);
-  let filtered = rankings.filter(
-    (r) => normalize(r.name).includes(searchTerm) || normalize(r.country).includes(searchTerm)
-  );
-
-  const selectedCountries = getSelectedCountries();
-  if (selectedCountries.length > 0) {
-    filtered = filtered.filter((r) => selectedCountries.includes(r.country));
-  }
-
-  const filterMode = getFilterMode();
-  const ownedSet = new Set(getOwnedGarages());
-  const displayRankings = filterMode === 'owned' ? filtered.filter((r) => ownedSet.has(r.id)) : filtered;
-  displayedRankings = displayRankings;
-
-  if (filterMode === 'owned' && displayRankings.length === 0) {
-    rankingsContent.innerHTML = `
-      <div class="empty-garages">
-        <p>No garages marked yet.</p>
-        <p class="hint">Click any city row, then click the star to mark it as your garage.</p>
-      </div>
-    `;
-    updateResultsCount(0, rankings.length);
-    return;
-  }
-
-  if (displayRankings.length === 0) {
-    let message: string;
-    if (searchTerm) {
-      const escaped = citySearch.value.trim().replace(/</g, '&lt;').replace(/>/g, '&gt;');
-      message = `No cities match '${escaped}'`;
-    } else if (selectedCountries.length > 0) {
-      message = 'No cities match your filters';
-    } else {
-      message = 'No results found';
-    }
-    rankingsContent.innerHTML = `
-      <div class="table-section">
-        <table class="table-rankings">
-          <thead>
-            <tr>
-              <th></th>
-              <th>#</th>
-              <th>City</th>
-              <th>Country</th>
-              <th class="tooltip" data-tooltip="Company facilities in this city">Depots</th>
-              <th class="tooltip" data-tooltip="Distinct cargo types available">Cargo</th>
-              <th class="tooltip" data-tooltip="Sum of top 5 body type EVs — fleet earning potential">Fleet EV</th>
-              <th class="tooltip" data-tooltip="Top earning trailer types for this city">Best Trailers</th>
-            </tr>
-          </thead>
-          <tbody>
-            <tr><td colspan="8" class="no-results" role="status">${message}</td></tr>
-          </tbody>
-        </table>
-      </div>
-    `;
-    updateGarageCount();
-    updateResultsCount(0, rankings.length);
-    return;
-  }
-
-  rankingsContent.innerHTML = `
-    <div class="table-section">
-      <h2>City Rankings (${displayRankings.length} cities)</h2>
-      <p class="table-hint">Ranked by combined fleet EV (top 5 trailer types). Click any city for details.</p>
-      <table class="table-rankings">
-        <thead>
-          <tr>
-            <th></th>
-            <th>#</th>
-            <th>City</th>
-            <th>Country</th>
-            <th class="tooltip" data-tooltip="Company facilities in this city">Depots</th>
-            <th class="tooltip" data-tooltip="Distinct cargo types available">Cargo</th>
-            <th class="tooltip" data-tooltip="Sum of top 5 body type EVs — fleet earning potential">Fleet EV</th>
-            <th class="tooltip" data-tooltip="Top earning trailer types for this city">Best Trailers</th>
-          </tr>
-        </thead>
-        <tbody>
-          ${displayRankings.map((r, i) => {
-            const trailerSummary = summarizeTrailers(r.topTrailers);
-            const starred = ownedSet.has(r.id);
-            const globalIndex = cachedRankings!.findIndex(cr => cr.id === r.id);
-            const tier = getScoreTier(globalIndex >= 0 ? globalIndex : i, cachedRankings!.length);
-            return `
-            <tr class="clickable${starred ? ' owned-garage' : ''}" data-city-id="${r.id}" tabindex="0">
-              <td class="garage-star" data-city-id="${r.id}" title="${starred ? 'Remove garage' : 'Mark as garage'}" tabindex="0" role="button" aria-label="${starred ? 'Remove garage for' : 'Toggle garage for'} ${r.name}">${starred ? '\u2605' : '\u2606'}</td>
-              <td>${i + 1}</td>
-              <td>${r.name}</td>
-              <td class="country">${r.country}</td>
-              <td>${r.depotCount}</td>
-              <td class="amount">${r.cargoTypes}</td>
-              <td class="score ${tier.className}" title="${tier.label}">${formatNumber(r.score)}${tier.label ? `<span class="score-tier-label">${tier.label.split(' — ')[0]}</span>` : ''}</td>
-              <td class="trailer-summary">${trailerSummary}</td>
-            </tr>
-          `;
-          }).join('')}
-        </tbody>
-      </table>
-    </div>
-  `;
-
-  // Star click/keyboard toggles garage without navigating to city
-  rankingsContent.querySelectorAll('.garage-star').forEach((star) => {
-    const toggleStar = (e: Event) => {
-      e.stopPropagation();
-      e.preventDefault();
-      const el = star as HTMLElement;
-      const cityId = el.dataset.cityId!;
-      const nowOwned = toggleOwnedGarage(cityId);
-      el.textContent = nowOwned ? '\u2605' : '\u2606';
-      el.title = nowOwned ? 'Remove garage' : 'Mark as garage';
-      const cityName = el.closest('tr')!.querySelector('td:nth-child(3)')!.textContent!;
-      el.setAttribute('aria-label', `${nowOwned ? 'Remove garage for' : 'Toggle garage for'} ${cityName}`);
-      const row = el.closest('tr')!;
-      row.classList.toggle('owned-garage', nowOwned);
-      updateGarageCount();
-    };
-    star.addEventListener('click', toggleStar);
-    star.addEventListener('keydown', (e) => {
-      if ((e as KeyboardEvent).key === 'Enter' || (e as KeyboardEvent).key === ' ') {
-        toggleStar(e);
-      }
-    });
-  });
-
-  rankingsContent.querySelectorAll('tr.clickable').forEach((row) => {
-    row.addEventListener('click', () => showCity((row as HTMLElement).dataset.cityId!));
-    row.addEventListener('keydown', (e) => {
-      if ((e as KeyboardEvent).key === 'Enter' || (e as KeyboardEvent).key === ' ') {
-        e.preventDefault();
-        showCity((row as HTMLElement).dataset.cityId!);
-      }
-    });
-  });
-
-  updateGarageCount();
-  updateResultsCount(displayRankings.length, rankings.length);
-}
-
-// ============================================
-// City detail rendering
-// ============================================
-
-async function ensureRankingsCached() {
-  if (cachedRankings === null && data && lookups) {
-    cachedRankings = await computeRankingsAsync(data, lookups);
-  }
-  if (displayedRankings === null && cachedRankings) {
-    // Apply current filters to build displayed rankings
-    const searchTerm = normalize(citySearch.value);
-    let filtered = cachedRankings.filter(
-      (r) => normalize(r.name).includes(searchTerm) || normalize(r.country).includes(searchTerm)
-    );
-    const selectedCountries = getSelectedCountries();
-    if (selectedCountries.length > 0) {
-      filtered = filtered.filter((r) => selectedCountries.includes(r.country));
-    }
-    const filterMode = getFilterMode();
-    const ownedSet = new Set(getOwnedGarages());
-    displayedRankings = filterMode === 'owned' ? filtered.filter((r) => ownedSet.has(r.id)) : filtered;
-  }
-}
-
-function renderFleetRow(entry: OptimalFleetEntry): string {
-  const countLabel = entry.count > 1 ? ` ×${entry.count}` : '';
-  const trailerLink = `trailers.html#body-${entry.bodyType}`;
-  return `
-    <tr>
-      <td>
-        <div><a href="${trailerLink}" class="body-type-link">${entry.displayName}${countLabel}</a></div>
-        <div class="trailer-spec">${entry.trailerSpec}</div>
-      </td>
-      <td class="amount">${formatNumber(entry.ev)}</td>
-      <td class="amount">${entry.cargoMatched}</td>
-    </tr>
-  `;
-}
-
-async function renderCity(cityId: string) {
-  await ensureRankingsCached();
-
-  const city = lookups!.citiesById.get(cityId);
-  if (!city) {
-    cityContent.innerHTML = '<div class="empty-state">City not found.</div>';
-    return;
-  }
-
-  const optimal = await computeFleetAsync(cityId, data!, lookups!);
-  if (!optimal) {
-    const emptyOwned = isOwnedGarage(cityId);
-    cityContent.innerHTML = `
-      <div class="city-header">
-        <div class="city-header-row">
-          <div>
-            <h2>${city.name}</h2>
-            <span class="country">${city.country}</span>
-          </div>
-          <button class="garage-toggle" id="city-garage-toggle"
-            aria-pressed="${emptyOwned}" aria-label="${emptyOwned ? 'Remove garage' : 'Mark as garage'}"
-            title="${emptyOwned ? 'Remove garage' : 'Mark as garage'}"
-            data-city-id="${cityId}">${emptyOwned ? '\u2605' : '\u2606'}</button>
-        </div>
-      </div>
-      <div class="empty-state">No cargo data for this city yet.</div>
-    `;
-    wireGarageToggle(cityId);
-    return;
-  }
-
-  const cityRank = getCityRank(cityId);
-  const cityCompanies = lookups!.cityCompanyMap.get(cityId) || [];
-  let depotCount = 0;
-  for (const { count } of cityCompanies) depotCount += count;
-
-  const rankingEntry = cachedRankings?.find(r => r.id === cityId);
-  const cargoTypes = rankingEntry?.cargoTypes ?? 0;
-  const score = rankingEntry?.score ?? 0;
-
-  // Compute score tier for city detail
-  let cityScoreTier: ScoreTier = { className: '', label: '' };
-  if (cachedRankings && rankingEntry) {
-    const rankIndex = cachedRankings.findIndex(r => r.id === cityId);
-    if (rankIndex >= 0) {
-      cityScoreTier = getScoreTier(rankIndex, cachedRankings.length);
-    }
-  }
-
-  const owned = isOwnedGarage(cityId);
-
-  cityContent.innerHTML = `
-    <div class="city-header">
-      <div class="city-header-row">
-        <div>
-          <h2>${city.name}</h2>
-          <span class="country">${city.country}</span>
-        </div>
-        <button class="garage-toggle" id="city-garage-toggle"
-          aria-pressed="${owned}" aria-label="${owned ? 'Remove garage' : 'Mark as garage'}"
-          title="${owned ? 'Remove garage' : 'Mark as garage'}"
-          data-city-id="${cityId}">${owned ? '\u2605' : '\u2606'}</button>
-      </div>
-    </div>
-
-    <div class="stats">
-      <div class="stat">
-        <div class="stat-value">${depotCount}</div>
-        <div class="stat-label">Depots</div>
-      </div>
-      <div class="stat">
-        <div class="stat-value">${cargoTypes}</div>
-        <div class="stat-label">Cargo Types</div>
-      </div>
-      <div class="stat">
-        <div class="stat-value">${cityRank ? formatRank(cityRank.rank, cityRank.total) : '-'}</div>
-        <div class="stat-label">Rank</div>
-      </div>
-      <div class="stat">
-        <div class="stat-value ${cityScoreTier.className}" title="${cityScoreTier.label}" aria-label="Score ${formatNumber(score)}, ${cityScoreTier.label || 'unranked'}">${formatNumber(score)}</div>
-        <div class="stat-label">Score${cityScoreTier.label ? ` — ${cityScoreTier.label.split(' — ')[0]}` : ''}</div>
-      </div>
-    </div>
-
-    <div class="table-section">
-      <div class="section-header">
-        <h2>Recommended Fleet — ${optimal.totalTrailers} trailers</h2>
-        <div class="export-buttons">
-          <button class="btn copy-btn" id="copy-fleet-btn" type="button">Copy Fleet</button>
-          <button class="btn export-btn" id="export-csv-btn" type="button">Export CSV</button>
-          <button class="btn export-btn" id="export-json-btn" type="button">Export JSON</button>
-        </div>
-      </div>
-      <table>
-        <thead>
-          <tr>
-            <th>Trailer Type</th>
-            <th class="tooltip" data-tooltip="Expected value per job cycle">EV</th>
-            <th class="tooltip" data-tooltip="Cargo types this trailer can haul">Cargo</th>
-          </tr>
-        </thead>
-        <tbody>
-          ${optimal.drivers.map(renderFleetRow).join('')}
-        </tbody>
-      </table>
-
-    </div>
-  `;
-
-  wireGarageToggle(cityId);
-  wireCopyFleetButton(city.name, optimal.drivers);
-  wireExportButtons(city.name, optimal.drivers, depotCount, cargoTypes, score);
-}
-
-function wireCopyFleetButton(cityName: string, drivers: OptimalFleetEntry[]) {
-  const copyBtn = document.getElementById('copy-fleet-btn') as HTMLButtonElement | null;
-  if (!copyBtn) return;
-
-  copyBtn.addEventListener('click', () => {
-    const lines = drivers.map(d => {
-      const countLabel = d.count > 1 ? ` x${d.count}` : '';
-      return `${d.displayName}${countLabel} (EV: ${formatNumber(d.ev)}, ${d.cargoMatched} cargo)`;
-    });
-    const text = `${cityName} Fleet:\n${lines.join('\n')}`;
-    copyToClipboard(text, copyBtn);
-  });
-}
-
-/**
- * Create a filesystem-safe filename from a string.
- * Transliterates diacritics to ASCII equivalents (e.g., ö→o, é→e)
- * and replaces only filesystem-unsafe characters.
- */
-function sanitizeFilename(name: string): string {
-  return name
-    .normalize('NFD')
-    .replace(/[\u0300-\u036f]/g, '')  // strip combining diacritics (ö→o, é→e, etc.)
-    .replace(/[/\\:*?"<>|]/g, '_')    // replace filesystem-unsafe chars
-    .replace(/_+/g, '_')              // collapse consecutive underscores
-    .replace(/^_|_$/g, '');           // trim leading/trailing underscores
-}
-
-function downloadFile(content: string, filename: string, mimeType: string): void {
-  const blob = new Blob([content], { type: mimeType });
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement('a');
-  a.href = url;
-  a.download = filename;
-  document.body.appendChild(a);
-  a.click();
-  document.body.removeChild(a);
-  URL.revokeObjectURL(url);
-}
-
-function exportToCSV(cityName: string, drivers: OptimalFleetEntry[]): void {
-  const headers = ['Trailer Type', 'Count', 'EV', 'Cargo Types'];
-  const rows = drivers.map(d => [
-    `"${d.displayName}"`,
-    d.count,
-    d.ev.toFixed(2),
-    d.cargoMatched,
-  ]);
-  const csv = [headers.join(','), ...rows.map(row => row.join(','))].join('\n');
-  const safeName = sanitizeFilename(cityName);
-  downloadFile(csv, `${safeName}_fleet.csv`, 'text/csv;charset=utf-8');
-}
-
-function exportToJSON(
-  cityName: string,
-  drivers: OptimalFleetEntry[],
-  depotCount: number,
-  cargoTypes: number,
-  score: number,
-): void {
-  const exportData = {
-    city: cityName,
-    exportedAt: new Date().toISOString(),
-    summary: {
-      depots: depotCount,
-      cargoTypes,
-      score,
-      totalTrailers: drivers.reduce((sum, d) => sum + d.count, 0),
-      trailerTypes: drivers.length,
-    },
-    fleet: drivers.map(d => ({
-      trailerType: d.displayName,
-      bodyType: d.bodyType,
-      count: d.count,
-      ev: d.ev,
-      cargoMatched: d.cargoMatched,
-    })),
+function debounce(fn: () => void, ms: number): () => void {
+  let timer: ReturnType<typeof setTimeout>;
+  return () => {
+    clearTimeout(timer);
+    timer = setTimeout(fn, ms);
   };
-  const json = JSON.stringify(exportData, null, 2);
-  const safeName = sanitizeFilename(cityName);
-  downloadFile(json, `${safeName}_fleet.json`, 'application/json');
 }
-
-function wireExportButtons(
-  cityName: string,
-  drivers: OptimalFleetEntry[],
-  depotCount: number,
-  cargoTypes: number,
-  score: number,
-): void {
-  document.getElementById('export-csv-btn')?.addEventListener('click', () => {
-    exportToCSV(cityName, drivers);
-  });
-  document.getElementById('export-json-btn')?.addEventListener('click', () => {
-    exportToJSON(cityName, drivers, depotCount, cargoTypes, score);
-  });
-}
-
-function wireGarageToggle(cityId: string) {
-  const garageToggle = document.getElementById('city-garage-toggle');
-  if (garageToggle) {
-    garageToggle.addEventListener('click', () => {
-      const nowOwned = toggleOwnedGarage(cityId);
-      garageToggle.textContent = nowOwned ? '\u2605' : '\u2606';
-      garageToggle.setAttribute('aria-pressed', String(nowOwned));
-      garageToggle.setAttribute('aria-label', nowOwned ? 'Remove garage' : 'Mark as garage');
-      garageToggle.title = nowOwned ? 'Remove garage' : 'Mark as garage';
-      // Sync the rankings table star if it exists
-      const rankingStar = rankingsContent.querySelector(`.garage-star[data-city-id="${cityId}"]`) as HTMLElement | null;
-      if (rankingStar) {
-        rankingStar.textContent = nowOwned ? '\u2605' : '\u2606';
-        rankingStar.title = nowOwned ? 'Remove garage' : 'Mark as garage';
-        const row = rankingStar.closest('tr')!;
-        row.classList.toggle('owned-garage', nowOwned);
-      }
-      updateGarageCount();
-    });
-  }
-}
-
 
 // ============================================
 // Navigation
@@ -667,7 +72,7 @@ async function showCity(cityId: string) {
   if (window.location.hash !== `#city-${cityId}`) {
     window.location.hash = `city-${cityId}`;
   }
-  await renderCity(cityId);
+  await renderCity(cityId, state, cityContent, rankingsContent, citySearch);
   window.scrollTo(0, 0);
 }
 
@@ -676,49 +81,20 @@ function showRankings() {
   cityView.style.display = 'none';
   rankingsView.style.display = 'block';
   window.location.hash = '';
-  renderRankings();
+  renderRankings(state, rankingsContent, citySearch, resultsCount, showCity);
 }
 
 function handleHashNavigation(): boolean {
   const hash = window.location.hash;
   if (hash.startsWith('#city-')) {
     const cityId = hash.replace('#city-', '');
-    if (cityId && lookups?.citiesById.has(cityId)) {
+    if (cityId && state.lookups?.citiesById.has(cityId)) {
       showCity(cityId);
       return true;
     }
   }
   return false;
 }
-
-// ============================================
-// Loading / Error states
-// ============================================
-
-function showLoading() {
-  rankingsContent.innerHTML = `
-    <div class="table-section" role="status" aria-label="Loading city data">
-      <h2>Loading city data...</h2>
-      <div class="skeleton-row"><div class="skeleton-cell narrow"></div><div class="skeleton-cell medium"></div><div class="skeleton-cell medium"></div><div class="skeleton-cell narrow"></div><div class="skeleton-cell narrow"></div><div class="skeleton-cell narrow"></div><div class="skeleton-cell medium"></div></div>
-      <div class="skeleton-row"><div class="skeleton-cell narrow"></div><div class="skeleton-cell medium"></div><div class="skeleton-cell medium"></div><div class="skeleton-cell narrow"></div><div class="skeleton-cell narrow"></div><div class="skeleton-cell narrow"></div><div class="skeleton-cell medium"></div></div>
-      <div class="skeleton-row"><div class="skeleton-cell narrow"></div><div class="skeleton-cell medium"></div><div class="skeleton-cell medium"></div><div class="skeleton-cell narrow"></div><div class="skeleton-cell narrow"></div><div class="skeleton-cell narrow"></div><div class="skeleton-cell medium"></div></div>
-    </div>
-  `;
-}
-
-function showError(errorMessage: string) {
-  rankingsContent.innerHTML = `
-    <div class="empty-state" role="alert" aria-live="assertive">
-      <p>Failed to load data</p>
-      <p class="error-detail">${errorMessage}</p>
-    </div>
-  `;
-}
-
-// ============================================
-// Settings
-// ============================================
-
 
 // ============================================
 // First-visit DLC banner
@@ -761,45 +137,42 @@ async function init() {
     if (collapsed) {
       onboardingSection.classList.add('collapsed');
       onboardingToggle.setAttribute('aria-expanded', 'false');
-      onboardingToggle.querySelector('.toggle-icon')!.textContent = '▶';
+      onboardingToggle.querySelector('.toggle-icon')!.textContent = '\u25b6';
     }
     onboardingToggle.addEventListener('click', () => {
       const isCollapsed = onboardingSection.classList.contains('collapsed');
       if (isCollapsed) {
         onboardingSection.classList.remove('collapsed');
         onboardingToggle.setAttribute('aria-expanded', 'true');
-        onboardingToggle.querySelector('.toggle-icon')!.textContent = '▼';
+        onboardingToggle.querySelector('.toggle-icon')!.textContent = '\u25bc';
         setOnboardingCollapsed(false);
       } else {
         onboardingSection.classList.add('collapsed');
         onboardingToggle.setAttribute('aria-expanded', 'false');
-        onboardingToggle.querySelector('.toggle-icon')!.textContent = '▶';
+        onboardingToggle.querySelector('.toggle-icon')!.textContent = '\u25b6';
         setOnboardingCollapsed(true);
       }
     });
   }
 
-  showLoading();
+  showLoading(rankingsContent);
 
   try {
     const page = await initPageData();
-    data = page.data;
-    lookups = page.lookups;
+    state.data = page.data;
+    state.lookups = page.lookups;
 
-    renderCountryCheckboxes();
-    updateCountryButtonText();
-
-    const savedFilterMode = getFilterMode();
-    filterToggle.querySelectorAll('.filter-btn').forEach((btn) => {
-      btn.classList.toggle('active', btn.getAttribute('data-filter') === savedFilterMode);
-    });
-    updateGarageCount();
+    // Initialize rankings view (filters, country dropdown, etc.)
+    initRankingsView(state, rankingsContent, citySearch, resultsCount, filterToggle, showCity);
 
     if (!handleHashNavigation()) {
-      renderRankings();
+      renderRankings(state, rankingsContent, citySearch, resultsCount, showCity);
     }
 
-    citySearch.addEventListener('input', debounce(renderRankings, 150));
+    citySearch.addEventListener('input', debounce(
+      () => renderRankings(state, rankingsContent, citySearch, resultsCount, showCity),
+      150,
+    ));
     backLink.addEventListener('click', showRankings);
     window.addEventListener('hashchange', () => {
       if (!handleHashNavigation()) showRankings();
@@ -813,55 +186,17 @@ async function init() {
         if (isCollapsed) {
           section.classList.remove('collapsed');
           howItWorksToggle.setAttribute('aria-expanded', 'true');
-          howItWorksToggle.querySelector('.toggle-icon')!.textContent = '▼';
+          howItWorksToggle.querySelector('.toggle-icon')!.textContent = '\u25bc';
         } else {
           section.classList.add('collapsed');
           howItWorksToggle.setAttribute('aria-expanded', 'false');
-          howItWorksToggle.querySelector('.toggle-icon')!.textContent = '▶';
+          howItWorksToggle.querySelector('.toggle-icon')!.textContent = '\u25b6';
         }
       });
     }
-
-    // Country filter dropdown
-    const countryFilterBtn = document.getElementById('country-filter-btn')!;
-    countryFilterBtn.addEventListener('click', (e) => { e.stopPropagation(); toggleDropdown(); });
-    countryFilterBtn.addEventListener('keydown', (e) => {
-      if ((e as KeyboardEvent).key === 'Enter' || (e as KeyboardEvent).key === ' ') {
-        e.preventDefault();
-        toggleDropdown();
-      }
-    });
-
-    document.addEventListener('keydown', (e) => {
-      const dropdown = document.getElementById('country-dropdown')!;
-      if (dropdown.style.display === 'none') return;
-      if ((e as KeyboardEvent).key === 'Escape') {
-        e.preventDefault();
-        closeDropdown();
-        countryFilterBtn.focus();
-      } else if ((e as KeyboardEvent).key === 'ArrowDown' || (e as KeyboardEvent).key === 'ArrowUp') {
-        e.preventDefault();
-        const checkboxes = Array.from(dropdown.querySelectorAll('input[type="checkbox"]'));
-        const currentIndex = checkboxes.findIndex(
-          (cb) => cb === document.activeElement || (cb as HTMLElement).parentElement === document.activeElement
-        );
-        const nextIndex = (e as KeyboardEvent).key === 'ArrowDown'
-          ? (currentIndex < checkboxes.length - 1 ? currentIndex + 1 : 0)
-          : (currentIndex > 0 ? currentIndex - 1 : checkboxes.length - 1);
-        (checkboxes[nextIndex] as HTMLElement).focus();
-      }
-    });
-
-    document.addEventListener('click', (e) => {
-      const dropdown = document.getElementById('country-dropdown')!;
-      const filterContainer = document.querySelector('.country-filter')!;
-      if (!filterContainer.contains(e.target as Node) && dropdown.style.display !== 'none') {
-        closeDropdown();
-      }
-    });
   } catch (err) {
     console.error('Failed to initialize:', err);
-    showError((err as Error).message || 'Unknown error occurred');
+    showError(rankingsContent, (err as Error).message || 'Unknown error occurred');
   }
 }
